refac: user-apiのフォルダ以降完了
