+++ src/api/user/factory.ts
@@ -1,10 +1,17 @@
 import { toUser } from "./model";
 import { userRepository } from "./repository";
-import { UpdateLoginUserParams } from "./types";
+import { LoginParams, SignupParams, UpdateLoginUserParams } from "./types";
 
-export const userFactory = () => {
+export const createFactory = () => {
   const repository = userRepository;
   return {
+    signup: async (params: SignupParams) => {
+      await repository.signup(params);
+    },
+    login: async (params: LoginParams) => {
+      await repository.login(params);
+    },
+    logout: async () => await repository.logout(),
     getLoginUser: async () => {
       const response = await repository.get();
       return toUser(response);
@@ -13,6 +20,5 @@ export const userFactory = () => {
       const response = await repository.update(params);
       return toUser(response);
     },
-    logout: async () => await repository.logout(),
   };
 };
