--- src/features/login/api/factory.ts
@@ -1,11 +1,0 @@
-import { userRepository } from "./repository";
-import { LoginParams } from "./types";
-
-export const userFactory = () => {
-  const repository = userRepository;
-  return {
-    login: async (params: LoginParams) => {
-      await repository.login(params);
-    },
-  };
-};
