--- src/features/login/api/types.ts
@@ -1,13 +1,0 @@
-export type LoginRequest = {
-  email: string;
-  password: string;
-};
-
-export type LoginResponse = {
-  status: string;
-};
-
-export type LoginParams = {
-  email: string;
-  password: string;
-};
