+++ src/features/profile/api/user/hooks/useQueryUser.ts
@@ -1,13 +1,14 @@
 import { useQuery } from "@tanstack/react-query";
-import { User } from "../model";
 import { FetchError } from "../../../../../pkg/api/util/fetchError";
-import { userFactory } from "../factory";
+import { User } from "../../../../../api/user/model";
+import { createFactory } from "../../../../../api/user/factory";
 
 export const useQueryLoginUser = () => {
+  const userFactory = createFactory();
   return useQuery<User, FetchError>({
     queryKey: ["user"],
     queryFn: async () => {
-      const user = await userFactory().getLoginUser();
+      const user = await userFactory.getLoginUser();
       return user;
     },
     staleTime: Infinity,
