--- src/features/profile/api/user/types.ts
@@ -1,19 +1,0 @@
-export type GetResponse = {
-  name: string;
-  email: string;
-  level: number;
-};
-
-export type UpdateLoginUserRequest = {
-  name: string;
-};
-
-export type UpdateLoginUserResponse = {
-  name: string;
-  email: string;
-  level: number;
-};
-
-export type UpdateLoginUserParams = {
-  name: string;
-};
