+++ src/features/signup/api/hooks/useMutateUser.ts
@@ -1,13 +1,15 @@
 import { useMutation } from "@tanstack/react-query";
-import { SignupParams } from "../types";
 import { FetchError } from "../../../../pkg/api/util/fetchError";
-import { userFactory } from "../factory";
 import { notifyFailed, notifySuccess } from "../../../../pkg/ui/toast";
+import { SignupParams } from "../../../../api/user/types";
+import { createFactory } from "../../../../api/user/factory";
 
 export const useMutateUser = () => {
+  const userFactory = createFactory();
+
   const signupMutation = useMutation({
     mutationFn: async (params: SignupParams) => {
-      return await userFactory().signup(params);
+      return await userFactory.signup(params);
     },
     onSuccess: () => {
       notifySuccess("サインアップに成功しました。");
