--- src/features/signup/api/types.ts
@@ -1,17 +1,0 @@
-export type SignupRequest = {
-  name: string;
-  email: string;
-  password: string;
-  passwordConfirmation: string;
-};
-
-export type SignupResponse = {
-  status: string;
-};
-
-export type SignupParams = {
-  name: string;
-  email: string;
-  password: string;
-  passwordConfirmation: string;
-};
